Reorganizing code and npm scripts.
